Add second one-sheet email blast and cal events
Now in addition to e-mailing staff the new freelancer e-mail it also
sends a one-sheet to the same recipients plus the freelancer with
important info for him. The original blast now also e-mails the producer
who filled out the form and includes a calendar event attachment for
easy importing into google and apple calendars.